:recycle: - refactor: refactor destruction list pages directory
