+++ frontend/src/pages/destructionlist/create/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListCreate";
