+++ frontend/src/pages/destructionlist/detail/Assignees.tsx
