+++ frontend/src/pages/destructionlist/detail/DestructionListDetail.css
