+++ frontend/src/pages/destructionlist/detail/constants.ts
