+++ frontend/src/pages/destructionlist/detail/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListDetail";
